🔨🤖 Fold poverty_inequality_file into inequality_comparison
keyvars had a single consumer (inequality_comparison) and no external use, so drop the
separate poverty_inequality_file step and build keyvars in-memory from the dimensional
datasets (shared.build_keyvars). Removes the indirection, the brittle series_code string
coupling, and the metadata round-trip. inequality_comparison output is byte-identical.
