+++ dag/poverty_inequality.yml
@@ -91,10 +91,6 @@ steps:
   # Inequality comparison data
   data://grapher/poverty_inequality/2025-01-22/inequality_comparison:
     - data://garden/poverty_inequality/2025-01-22/inequality_comparison:
-      # Poverty and inequality file for inequality comparisons at OWID
-      - data://garden/poverty_inequality/2025-01-22/poverty_inequality_file:
-        - data://garden/wb/2026-03-24/world_bank_pip
-        - data://garden/wid/2026-02-10/world_inequality_database
       - data://garden/demography/2024-07-15/population
       - data://garden/regions/2023-01-01/regions
       - data://garden/wb/2026-03-24/world_bank_pip
