--- etl/steps/data/garden/poverty_inequality/2025-01-22/poverty_inequality_file.meta.yml
@@ -1,3 +1,0 @@
-dataset:
-  owners:
-    - Pablo Arriagada
